+++ product_code/user_service/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [v.0.9.2](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.9.1...user-v0.9.2) (2020-09-28)
+## Feature
+ - Added Default Profile Picture
+## Refactor
+## Bugfix
+ - PUT of a Profile Picture returns now a 200
+
 # [v.0.9.1](https://github.com/upb-uc4/University-Credits-4.0/compare/user-v0.8.3...user-v0.9.1) (2020-09-28)
 ## Feature
  - Added Support of Profile Pictures